requirements.txt
requests
python-telegram-bot
chromadb
sentence_transformers